feature/ui-testing Refactor:Move MockWebServer to idlingResources flavor
